% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{aupr}
\alias{aupr}
\title{Compute Area Under Prediction Recall Curve.}
\usage{
aupr(pred, obs)
}
\arguments{
\item{pred}{numeric vector of predcted scores.}

\item{obs}{numeric vector of true labels (0 and 1).}
}
\value{
numeric value of Area Under Precision Recall Curve.
}
\description{
A function to compute AUPRC.
}
\keyword{BindSpace}
