% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{eval_peak}
\alias{eval_peak}
\title{Evaluate on ChIP-seq / ATAC-seq peaks.}
\usage{
eval_peak(input, model, TF = NULL)
}
\arguments{
\item{input}{DNAStringSet object of query sequences (all sequences must be of the same length and don't have N).}

\item{model}{list object of embedding model. Contains four elements: kmer_embed, label_embed, kmer2kmerwc and TF 0.05 empirical thresholds}

\item{TF}{a character value. If a particular TF is specsified, a full score matrix will be returned for this TF. (A matrix of dimension #peaks * #windows).}
}
\value{
a list of two elements. 
First element is a matrix of size #peaks * #labels. For each peak, a single score is assigned for each label, which is the max score across all 20bp windows.
Second element is a binary thresholded matrix of first.
Third element is not Null only when TF is specified. It is a matrix of size #peaks * #windows, that is the score matrix for specified TF.
}
\description{
A function to predict binding affinity to peaks. A score will be computed for every 20bp window of the peak.
}
\keyword{BindSpace}
