% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{eval_probe}
\alias{eval_probe}
\title{Evaluate on probe sequence.}
\usage{
eval_probe(input, model)
}
\arguments{
\item{input}{DNAStringSet object of query sequences.}

\item{model}{list object of embedding model. Contains four elements: kmer_embed, label_embed, kmer2kmerwc and TF 0.05 empirical thresholds}
}
\value{
a matrix of predicted specificity for all sequences and all labels.
}
\description{
A function to predict specificity of 20bp sequences.
}
\keyword{BindSpace}
