% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{F1}
\alias{F1}
\title{Compute F1 score.}
\usage{
F1(pred, obs)
}
\arguments{
\item{pred}{numeric vector of predcted labels (0 and 1).}

\item{obs}{numeric vector of true labels (0 and 1).}
}
\value{
numeric vector of precision, recall, and F1 score.
}
\description{
A function to compute F1 score.
}
\keyword{BindSpace}
