% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get.time}
\alias{get.time}
\title{Get current system time.}
\usage{
get.time()
}
\value{
None.
}
\description{
A function to get current system time. This is the same function as from SeqGL.
}
\keyword{BindSpace}
