% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{makeFeats}
\alias{makeFeats}
\title{Compute kmer frequency.}
\usage{
makeFeats(seqs, k, rc = F)
}
\arguments{
\item{seqs}{DNAStringSet object. query sequences.}

\item{k}{int value. kmer length.}

\item{rc}{logical value. whether reverseComplement sequence should also be counted.}
}
\value{
a kmer frequency table.
}
\description{
A function to predict specificity of 20bp sequences.
}
\keyword{BindSpace}
