% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plot binding score Heatmap.}
\usage{
plotHeatmap(m)
}
\arguments{
\item{m}{matrix of per window per peak binding score for a given TF}
}
\value{
a ComplexHeatmap object
}
\description{
A function to plot binding score.
}
\keyword{BindSpace}
